+++ IslandSFML/UI.h
@@ -6,7 +6,7 @@ class UI {
 private:
 	sf::RenderWindow *window;
 	std::vector<std::unique_ptr<sf::RectangleShape>> monitors;
-	void createMonitor(int _posX, int _posY, int _sizeX, int _sizeY, sf::Color _fColor, sf::Color _oColor, int _thickness);
+	void createMonitor(float _posX, float _posY, float _sizeX, float _sizeY, sf::Color _fColor, sf::Color _oColor, float _thickness);
 public:
 	void drawUI();
 	UI(sf::RenderWindow *window);
